feat:refactor some code
