+++ lib/src/extension/color_extension.dart
@@ -1,0 +1,51 @@
+import 'package:pdf/pdf.dart';
+
+extension ColorExtension on PdfColor {
+  /// Try to parse the `rgba(red, greed, blue, alpha)`
+  /// from the string.
+  static PdfColor? tryFromRgbaString(String colorString) {
+    final reg = RegExp(r'rgba\((\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*\)');
+    final match = reg.firstMatch(colorString);
+    if (match == null) {
+      return null;
+    }
+
+    if (match.groupCount < 4) {
+      return null;
+    }
+    final redStr = match.group(1);
+    final greenStr = match.group(2);
+    final blueStr = match.group(3);
+    final alphaStr = match.group(4);
+
+    final red = redStr != null ? int.tryParse(redStr) : null;
+    final green = greenStr != null ? int.tryParse(greenStr) : null;
+    final blue = blueStr != null ? int.tryParse(blueStr) : null;
+    final alpha = alphaStr != null ? int.tryParse(alphaStr) : null;
+
+    if (red == null || green == null || blue == null || alpha == null) {
+      return null;
+    }
+
+    return PdfColor.fromInt(
+        hexOfRGBA(red, green, blue, opacity: alpha.toDouble()));
+  }
+
+  String toRgbaString() {
+    return 'rgba($red, $green, $blue, $alpha)';
+  }
+}
+
+int hexOfRGBA(int r, int g, int b, {double opacity = 1}) {
+  r = (r < 0) ? -r : r;
+  g = (g < 0) ? -g : g;
+  b = (b < 0) ? -b : b;
+  opacity = (opacity < 0) ? -opacity : opacity;
+  opacity = (opacity > 1) ? 255 : opacity * 255;
+  r = (r > 255) ? 255 : r;
+  g = (g > 255) ? 255 : g;
+  b = (b > 255) ? 255 : b;
+  int a = opacity.toInt();
+  return int.parse(
+      '0x${a.toRadixString(16)}${r.toRadixString(16)}${g.toRadixString(16)}${b.toRadixString(16)}');
+}
