+++ lib/src/html_to_widgets.dart
@@ -4,48 +4,62 @@ import 'dart:typed_data';
 import 'package:html/dom.dart' as dom;
 import 'package:html/parser.dart' show parse;
 import 'package:htmltopdfwidgets/src/attributes.dart';
+import 'package:htmltopdfwidgets/src/extension/int_extensions.dart';
 import 'package:http/http.dart';
 
 import '../htmltopdfwidgets.dart';
+import 'extension/color_extension.dart';
+import 'html_tags.dart';
+import 'pdfwidgets/bullet_list.dart';
+import 'pdfwidgets/number_list.dart';
 
+//html deocoder that deocde html and convert it into pdf widgets
 class WidgetsHTMLDecoder {
-  final Font? font;
-  final HtmlTagStyle customStyles;
-  final List<Font> fontFallback;
-  const WidgetsHTMLDecoder(
-      {this.font,
-      required this.fontFallback,
-      this.customStyles = const HtmlTagStyle()});
+  //default font font the pdf if it not provided custo
+  // Constructor for the HTML decoder
+  final Font? font; // Font for the PDF, if not provided, use default
+  final HtmlTagStyle customStyles; // Custom styles for HTML tags
+  final List<Font> fontFallback; // Fallback fonts
+  const WidgetsHTMLDecoder({
+    this.font,
+    required this.fontFallback,
+    this.customStyles = const HtmlTagStyle(),
+  });
 
-  Future<List<Widget>> convert(
-    String html,
-  ) async {
+  /// The class takes an HTML string as input and returns a list of Widgets. The Widgets
+  /// are created based on the tags and attributes in the HTML string.
+  Future<List<Widget>> convert(String html) async {
+    // Parse the HTML document using the html package
     final document = parse(html);
     final body = document.body;
     if (body == null) {
       return [];
     }
-    List<Widget> nodes = await _parseElement(
-      body.nodes,
-    );
+    // Call the private _parseElement function to process the HTML nodes
+    List<Widget> nodes = await _parseElement(body.nodes);
 
     return nodes;
   }
+
+  /// Converts the given HTML string to a list of Widgets.
+  /// and returns the list of widgets
 
   Future<List<Widget>> _parseElement(
     Iterable<dom.Node> domNodes,
   ) async {
     final List<Widget> delta = [];
     final result = <Widget>[];
+    //find dom node in and check if its element or not than convert it according to its specs
     for (final domNode in domNodes) {
       if (domNode is dom.Element) {
         final localName = domNode.localName;
-
+        // Check if the element is a simple formatting element like <span>, <bold>, or <italic>
         if (HTMLTags.formattingElements.contains(localName)) {
           final attributes = await _parserFormattingElementAttributes(domNode);
 
           result.add(Text(domNode.text, style: attributes));
         } else if (HTMLTags.specialElements.contains(localName)) {
+          // Handle special elements (e.g., headings, lists, images)
           result.addAll(
             await _parseSpecialElements(
               domNode,
@@ -54,53 +68,70 @@ class WidgetsHTMLDecoder {
           );
         }
       } else if (domNode is dom.Text) {
+        // Process text nodes and add them to delta
         delta.add(Text(domNode.text,
             style: TextStyle(font: font, fontFallback: fontFallback)));
       } else {
         assert(false, 'Unknown node type: $domNode');
       }
     }
+    // If there are text nodes in delta, wrap them in a Wrap widget and add to the result
     if (delta.isNotEmpty) {
       result.add(Wrap(children: delta));
     }
     return result;
   }
 
+  // Function to parse special HTML elements (e.g., headings, lists, images)
   Future<Iterable<Widget>> _parseSpecialElements(
     dom.Element element, {
     required String type,
   }) async {
     final localName = element.localName;
     switch (localName) {
+      // Handle heading level 1
       case HTMLTags.h1:
         return [await _parseHeadingElement(element, level: 1)];
+      // Handle heading level 2
       case HTMLTags.h2:
         return [await _parseHeadingElement(element, level: 2)];
+      // Handle heading level 3
       case HTMLTags.h3:
         return [await _parseHeadingElement(element, level: 3)];
+      // Handle heading level 4
       case HTMLTags.h4:
         return [await _parseHeadingElement(element, level: 4)];
+      // Handle heading level 5
       case HTMLTags.h5:
         return [await _parseHeadingElement(element, level: 5)];
+      // Handle heading level 6
       case HTMLTags.h6:
         return [await _parseHeadingElement(element, level: 6)];
+      // Handle unorder list
       case HTMLTags.unorderedList:
         return await _parseUnOrderListElement(element);
+      // Handle ordered list and converts its childrens to widgets
       case HTMLTags.orderedList:
         return await _parseOrderListElement(element);
+      //if simple list is found it will handle accoridingly
       case HTMLTags.list:
         return await _parseListElement(
           element,
           type: type,
         );
+      // it handles the simple paragraph element
       case HTMLTags.paragraph:
         return [await _parseParagraphElement(element)];
+      // Handle block quote tag
       case HTMLTags.blockQuote:
         return await _parseBlockQuoteElement(element);
+      // Handle the image tag
       case HTMLTags.image:
         return [await _parseImageElement(element)];
+      // Handle the line break tag
       case HTMLTags.br:
         return [Text("\n")];
+      // if no special element is found it treated as simple parahgraph
       default:
         return [await _parseParagraphElement(element)];
     }
@@ -110,6 +141,7 @@ class WidgetsHTMLDecoder {
     return Text(text);
   }
 
+  /// Parses the attributes of a formatting element and returns a TextStyle.
   Future<TextStyle> _parserFormattingElementAttributes(
       dom.Element element) async {
     final localName = element.localName;
@@ -117,30 +149,36 @@ class WidgetsHTMLDecoder {
     TextStyle attributes = TextStyle(fontFallback: fontFallback, font: font);
     final List<TextDecoration> decoration = [];
     switch (localName) {
+      // Handle <bold> element
       case HTMLTags.bold:
         attributes = attributes.copyWith(fontWeight: FontWeight.bold)
           ..merge(customStyles.boldStyle);
         break;
+      // Handle <strong> element
       case HTMLTags.strong:
         attributes = attributes.copyWith(fontWeight: FontWeight.bold)
           ..merge(customStyles.boldStyle);
         break;
+      // Handle <em> element
       case HTMLTags.em:
         attributes = attributes.copyWith(fontStyle: FontStyle.italic)
           ..merge(customStyles.italicStyle);
         break;
+      // Handle <italic> element
       case HTMLTags.italic:
         attributes = attributes.copyWith(fontStyle: FontStyle.italic)
           ..merge(customStyles.italicStyle);
         break;
+      // Handle <u> element
       case HTMLTags.underline:
         decoration.add(TextDecoration.underline);
         break;
+      // Handle <del> element
       case HTMLTags.del:
         decoration.add(TextDecoration.lineThrough);
 
         break;
-
+      // Handle <span> element
       case HTMLTags.span:
         final deltaAttributes = _getDeltaAttributesFromHtmlAttributes(
           element.attributes,
@@ -150,6 +188,7 @@ class WidgetsHTMLDecoder {
           decoration.add(deltaAttributes.decoration!);
         }
         break;
+      // Handle <a> element
       case HTMLTags.anchor:
         final href = element.attributes['href'];
         if (href != null) {
@@ -160,9 +199,11 @@ class WidgetsHTMLDecoder {
             ..merge(customStyles.linkStyle);
         }
         break;
+      // Handle <p> element for additional safety
       case HTMLTags.paragraph:
         attributes = attributes..merge(customStyles.paragraphStyle);
         break;
+      // Handle <code> element
       case HTMLTags.code:
         attributes = attributes.copyWith(
             background: const BoxDecoration(color: PdfColors.red))
@@ -183,6 +224,7 @@ class WidgetsHTMLDecoder {
     return attributes.copyWith(decoration: TextDecoration.combine(decoration));
   }
 
+  // Function to parse a heading element and return a RichText widget
   Future<Widget> _parseHeadingElement(
     dom.Element element, {
     required int level,
@@ -199,54 +241,16 @@ class WidgetsHTMLDecoder {
             style: TextStyle(font: font, fontFallback: fontFallback)));
       }
     }
+    // Return a RichText widget with the parsed text and styles
     return RichText(
         text: TextSpan(
             children: delta,
             style: TextStyle(
-                fontSize: getHeadingSize(level), fontWeight: FontWeight.bold)
-              ..merge(getHeadingStyle(level))));
+                fontSize: level.getHeadingSize, fontWeight: FontWeight.bold)
+              ..merge(level.getHeadingStyle(customStyles))));
   }
 
-  static double getHeadingSize(int level) {
-    switch (level) {
-      case 1:
-        return 32;
-      case 2:
-        return 28;
-      case 3:
-        return 20;
-      case 4:
-        return 17;
-      case 5:
-        return 14;
-      case 6:
-        return 10;
-      default:
-        return 32;
-    }
-  }
-
-  TextStyle? getHeadingStyle(
-    int level,
-  ) {
-    switch (level) {
-      case 1:
-        return customStyles.h1Style;
-      case 2:
-        return customStyles.h2Style;
-      case 3:
-        return customStyles.h3Style;
-      case 4:
-        return customStyles.h4Style;
-      case 5:
-        return customStyles.h5Style;
-      case 6:
-        return customStyles.h6Style;
-      default:
-        return customStyles.h1Style;
-    }
-  }
-
+// Function to parse a block quote element and return a list of widgets
   Future<List<Widget>> _parseBlockQuoteElement(dom.Element element) async {
     final result = <Widget>[];
     for (final child in element.children) {
@@ -256,6 +260,7 @@ class WidgetsHTMLDecoder {
     return result;
   }
 
+// Function to parse an unordered list element and return a list of widgets
   Future<Iterable<Widget>> _parseUnOrderListElement(dom.Element element) async {
     final result = <Widget>[];
     for (final child in element.children) {
@@ -265,6 +270,7 @@ class WidgetsHTMLDecoder {
     return result;
   }
 
+  // Function to parse an ordered list element and return a list of widgets
   Future<Iterable<Widget>> _parseOrderListElement(dom.Element element) async {
     final result = <Widget>[];
     for (var i = 0; i < element.children.length; i++) {
@@ -275,16 +281,26 @@ class WidgetsHTMLDecoder {
     return result;
   }
 
+  // Function to parse a list element (unordered or ordered) and return a list of widgets
   Future<Iterable<Widget>> _parseListElement(
     dom.Element element, {
     required String type,
     int? index,
   }) async {
     final delta = await _parseDeltaElement(element);
+    // Build a bullet list widget
     if (type == BuiltInAttributeKey.bulletedList) {
-      return [buildBulletwidget(delta)];
+      return [buildBulletwidget(delta, customStyles: customStyles)];
+      // Build a numbered list widget
     } else if (type == BuiltInAttributeKey.numberList) {
-      return [buildNumberwdget(delta, index: index!)];
+      return [
+        buildNumberwdget(delta,
+            index: index!,
+            customStyles: customStyles,
+            font: font,
+            fontFallback: fontFallback)
+      ];
+      // Build a quote  widget
     } else if (type == BuiltInAttributeKey.quote) {
       return [buildQuotewidget(delta)];
     } else {
@@ -292,11 +308,13 @@ class WidgetsHTMLDecoder {
     }
   }
 
+  // Function to parse a paragraph element and return a widget
   Future<Widget> _parseParagraphElement(dom.Element element) async {
     final delta = await _parseDeltaElement(element);
     return delta;
   }
 
+// Function to parse an image element and download image as bytes  and return an Image widget
   Future<Widget> _parseImageElement(dom.Element element) async {
     final src = element.attributes["src"];
     try {
@@ -312,6 +330,7 @@ class WidgetsHTMLDecoder {
     }
   }
 
+// Function to download and save an image from a URL
   Future<Uint8List> _saveImage(String url) async {
     try {
       // Download image
@@ -325,15 +344,18 @@ class WidgetsHTMLDecoder {
     }
   }
 
+  // Function to parse a complex HTML element and return a widget
   Future<Widget> _parseDeltaElement(dom.Element element) async {
     final delta = <Widget>[];
     final children = element.nodes.toList();
     final childNodes = <Widget>[];
     for (final child in children) {
+      // Recursively parse child elements
       if (child is dom.Element) {
         if (child.children.isNotEmpty) {
           childNodes.addAll(await _parseElement(child.children));
         } else {
+          // Handle special elements (e.g., headings, lists) within a paragraph
           if (HTMLTags.specialElements.contains(child.localName)) {
             childNodes.addAll(
               await _parseSpecialElements(
@@ -342,28 +364,32 @@ class WidgetsHTMLDecoder {
               ),
             );
           } else {
+            // Parse text and attributes within the paragraph
             final attributes = await _parserFormattingElementAttributes(child)
               ..merge(customStyles.paragraphStyle);
             delta.add(Text(child.text, style: attributes));
           }
         }
       } else {
+        // Process text nodes and add them to delta variable
         delta.add(Text(child.text ?? "",
             style: TextStyle(font: font, fontFallback: fontFallback)
               ..merge(customStyles.paragraphStyle)));
       }
     }
+    // Create a column with wrapped text and child nodes
     return Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [Wrap(children: delta), ...childNodes]);
   }
 
+  // Utility function to convert a CSS string to a map of CSS properties
   static Map<String, String> _cssStringToMap(String? cssString) {
     final result = <String, String>{};
     if (cssString == null) {
       return result;
     }
-
+// Split the CSS string into key-value pairs and add them to the result map
     final entries = cssString.split(";");
     for (final entry in entries) {
       final tuples = entry.split(":");
@@ -376,9 +402,11 @@ class WidgetsHTMLDecoder {
     return result;
   }
 
+// Function to extract text styles from HTML attributes
   TextStyle _getDeltaAttributesFromHtmlAttributes(
       LinkedHashMap<Object, String> htmlAttributes) {
     TextStyle style = const TextStyle();
+    //extract styls from the inline css
     final styleString = htmlAttributes["style"];
     final cssMap = _cssStringToMap(styleString);
 
@@ -431,33 +459,9 @@ class WidgetsHTMLDecoder {
     return style.copyWith(decoration: TextDecoration.combine(textdecorations));
   }
 
-  Widget defaultIndex(int index) {
-    return Container(
-      width: 20,
-      padding: const EdgeInsets.only(right: 5.0),
-      child: Text('$index.',
-          style: TextStyle(
-            font: font,
-            fontFallback: fontFallback,
-          )..merge(customStyles.listIndexStyle)),
-    );
-  }
+//return the number list child with its current number
 
-  Widget buildNumberwdget(Widget childValue, {required int index}) {
-    Widget child = Container(
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisAlignment: MainAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          defaultIndex(index),
-          Flexible(child: childValue),
-        ],
-      ),
-    );
-    return child;
-  }
-
+//return the quote widget
   Widget buildQuotewidget(
     Widget childValue,
   ) {
@@ -478,163 +482,6 @@ class WidgetsHTMLDecoder {
     );
     return child;
   }
-
-  Widget buildBulletwidget(Widget childValue) {
-    Widget child = Container(
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisAlignment: MainAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          _BulletedListIcon(style: customStyles),
-          Flexible(child: childValue),
-        ],
-      ),
-    );
-    return child;
-  }
 }
-
-class _BulletedListIcon extends StatelessWidget {
-  final HtmlTagStyle style;
-  _BulletedListIcon({required this.style});
-
-  @override
-  Widget build(Context context) {
-    return SizedBox(
-      width: 22,
-      height: 22,
-      child: Padding(
-        padding: const EdgeInsets.only(right: 5.0),
-        child: Center(
-            child: Container(
-                width: 5,
-                height: 5,
-                decoration: BoxDecoration(
-                    shape: BoxShape.circle,
-                    color: style.bulletListIconColor ?? PdfColors.black))),
-      ),
-    );
-  }
-}
-
-class HTMLTags {
-  static const h1 = 'h1';
-  static const h2 = 'h2';
-  static const h3 = 'h3';
-  static const h4 = 'h4';
-  static const h5 = 'h5';
-  static const h6 = 'h6';
-  static const orderedList = 'ol';
-  static const unorderedList = 'ul';
-  static const list = 'li';
-  static const paragraph = 'p';
-  static const image = 'img';
-  static const anchor = 'a';
-  static const italic = 'i';
-  static const em = 'em';
-  static const bold = 'b';
-  static const br = 'br';
-  static const underline = 'u';
-  static const del = 'del';
-  static const strong = 'strong';
-  static const checkbox = 'input';
-  static const span = 'span';
-  static const code = 'code';
-  static const blockQuote = 'blockquote';
-  static const div = 'div';
-  static const divider = 'hr';
-
-  static List<String> formattingElements = [
-    HTMLTags.anchor,
-    HTMLTags.italic,
-    HTMLTags.em,
-    HTMLTags.bold,
-    HTMLTags.underline,
-    HTMLTags.del,
-    HTMLTags.strong,
-    HTMLTags.span,
-    HTMLTags.code,
-  ];
-
-  static List<String> specialElements = [
-    HTMLTags.h1,
-    HTMLTags.h2,
-    HTMLTags.h3,
-    HTMLTags.h4,
-    HTMLTags.h5,
-    HTMLTags.h6,
-    HTMLTags.div,
-    HTMLTags.br,
-    HTMLTags.unorderedList,
-    HTMLTags.orderedList,
-    HTMLTags.list,
-    HTMLTags.paragraph,
-    HTMLTags.blockQuote,
-    HTMLTags.checkbox,
-    HTMLTags.image
-  ];
-
-  static bool isTopLevel(String tag) {
-    return tag == h1 ||
-        tag == h2 ||
-        tag == h3 ||
-        tag == h4 ||
-        tag == h5 ||
-        tag == h6 ||
-        tag == checkbox ||
-        tag == paragraph ||
-        tag == div ||
-        tag == blockQuote;
-  }
-}
-
-extension ColorExtension on PdfColor {
-  /// Try to parse the `rgba(red, greed, blue, alpha)`
-  /// from the string.
-  static PdfColor? tryFromRgbaString(String colorString) {
-    final reg = RegExp(r'rgba\((\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*\)');
-    final match = reg.firstMatch(colorString);
-    if (match == null) {
-      return null;
-    }
-
-    if (match.groupCount < 4) {
-      return null;
-    }
-    final redStr = match.group(1);
-    final greenStr = match.group(2);
-    final blueStr = match.group(3);
-    final alphaStr = match.group(4);
-
-    final red = redStr != null ? int.tryParse(redStr) : null;
-    final green = greenStr != null ? int.tryParse(greenStr) : null;
-    final blue = blueStr != null ? int.tryParse(blueStr) : null;
-    final alpha = alphaStr != null ? int.tryParse(alphaStr) : null;
-
-    if (red == null || green == null || blue == null || alpha == null) {
-      return null;
-    }
-
-    return PdfColor.fromInt(
-        hexOfRGBA(red, green, blue, opacity: alpha.toDouble()));
-  }
-
-  String toRgbaString() {
-    return 'rgba($red, $green, $blue, $alpha)';
-  }
-}
-
-int hexOfRGBA(int r, int g, int b, {double opacity = 1}) {
-  r = (r < 0) ? -r : r;
-  g = (g < 0) ? -g : g;
-  b = (b < 0) ? -b : b;
-  opacity = (opacity < 0) ? -opacity : opacity;
-  opacity = (opacity > 1) ? 255 : opacity * 255;
-  r = (r > 255) ? 255 : r;
-  g = (g > 255) ? 255 : g;
-  b = (b > 255) ? 255 : b;
-  int a = opacity.toInt();
-  return int.parse(
-      '0x${a.toRadixString(16)}${r.toRadixString(16)}${g.toRadixString(16)}${b.toRadixString(16)}');
-}
+//return bullet widget with childreen
+
